+++ benchmark/src/main/java/io/github/pr0methean/betterrandom/benchmark/ReseedingSplittableRandomAdapterBenchmark.java
@@ -5,13 +5,13 @@ import io.github.pr0methean.betterrandom.seed.RandomSeederThread;
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import java.util.Random;
 
-import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.SECURE_RANDOM_SEED_GENERATOR;
+import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.DEFAULT_INSTANCE;
 
 public class ReseedingSplittableRandomAdapterBenchmark extends AbstractRandomBenchmark {
 
-  private final RandomSeederThread thread = new RandomSeederThread(SECURE_RANDOM_SEED_GENERATOR);
+  private final RandomSeederThread thread = new RandomSeederThread(DEFAULT_INSTANCE);
 
   @Override protected Random createPrng() throws SeedException {
-    return ReseedingSplittableRandomAdapter.getInstance(thread, SECURE_RANDOM_SEED_GENERATOR);
+    return ReseedingSplittableRandomAdapter.getInstance(thread, DEFAULT_INSTANCE);
   }
 }