+++ benchmark/src/main/java/io/github/pr0methean/betterrandom/benchmark/ReseedingThreadLocalRandomWrapperAesCounterRandom128Benchmark.java
@@ -4,13 +4,13 @@ import io.github.pr0methean.betterrandom.prng.AesCounterRandom;
 import io.github.pr0methean.betterrandom.prng.concurrent.ReseedingThreadLocalRandomWrapper;
 import java.util.Random;
 
-import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.SECURE_RANDOM_SEED_GENERATOR;
+import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.DEFAULT_INSTANCE;
 
 public class ReseedingThreadLocalRandomWrapperAesCounterRandom128Benchmark
     extends AbstractRandomBenchmark {
 
   @Override protected Random createPrng() {
-    return new ReseedingThreadLocalRandomWrapper(SECURE_RANDOM_SEED_GENERATOR,
-        () -> new AesCounterRandom(SECURE_RANDOM_SEED_GENERATOR.generateSeed(16)));
+    return new ReseedingThreadLocalRandomWrapper(DEFAULT_INSTANCE,
+        () -> new AesCounterRandom(DEFAULT_INSTANCE.generateSeed(16)));
   }
 }