+++ benchmark/src/main/java/io/github/pr0methean/betterrandom/benchmark/ThreadLocalRandomWrapperAesCounterRandom128Benchmark.java
@@ -5,12 +5,12 @@ import io.github.pr0methean.betterrandom.prng.concurrent.ThreadLocalRandomWrappe
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import java.util.Random;
 
-import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.SECURE_RANDOM_SEED_GENERATOR;
+import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.DEFAULT_INSTANCE;
 
 public class ThreadLocalRandomWrapperAesCounterRandom128Benchmark extends AbstractRandomBenchmark {
 
   @Override protected Random createPrng() throws SeedException {
-    return new ThreadLocalRandomWrapper(16, SECURE_RANDOM_SEED_GENERATOR,
+    return new ThreadLocalRandomWrapper(16, DEFAULT_INSTANCE,
         AesCounterRandom::new);
   }
 }