+++ benchmark/src/main/java/io/github/pr0methean/betterrandom/benchmark/XorShiftRandomBenchmark.java
@@ -4,11 +4,11 @@ import io.github.pr0methean.betterrandom.prng.XorShiftRandom;
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import java.util.Random;
 
-import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.SECURE_RANDOM_SEED_GENERATOR;
+import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.DEFAULT_INSTANCE;
 
 public class XorShiftRandomBenchmark extends AbstractRandomBenchmarkWithReseeding {
 
   @Override protected Random createPrng() throws SeedException {
-    return new XorShiftRandom(SECURE_RANDOM_SEED_GENERATOR);
+    return new XorShiftRandom(DEFAULT_INSTANCE);
   }
 }