+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/AesCounterRandomDemo.java
@@ -1,6 +1,6 @@
 package io.github.pr0methean.betterrandom.prng;
 
-import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.SECURE_RANDOM_SEED_GENERATOR;
+import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.DEFAULT_INSTANCE;
 
 import io.github.pr0methean.betterrandom.seed.RandomSeederThread;
 import io.github.pr0methean.betterrandom.seed.SeedException;
@@ -10,8 +10,8 @@ public enum AesCounterRandomDemo {
   ;
 
   public static void main(final String[] args) throws SeedException {
-    final AesCounterRandom random = new AesCounterRandom(SECURE_RANDOM_SEED_GENERATOR);
-    new RandomSeederThread(SECURE_RANDOM_SEED_GENERATOR).add(random);
+    final AesCounterRandom random = new AesCounterRandom(DEFAULT_INSTANCE);
+    new RandomSeederThread(DEFAULT_INSTANCE).add(random);
     final byte[] randomBytes = new byte[32];
     for (int i = 0; i < 20; i++) {
       random.nextBytes(randomBytes);