+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/concurrent/ReseedingSplittableRandomAdapterDemo.java
@@ -1,6 +1,6 @@
 package io.github.pr0methean.betterrandom.prng.concurrent;
 
-import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.SECURE_RANDOM_SEED_GENERATOR;
+import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.DEFAULT_INSTANCE;
 
 import io.github.pr0methean.betterrandom.seed.RandomSeederThread;
 import io.github.pr0methean.betterrandom.seed.SeedException;
@@ -30,9 +30,9 @@ public enum ReseedingSplittableRandomAdapterDemo {
         ThreadLocal.withInitial(() -> Arrays.asList(cards.clone()));
     final ScheduledThreadPoolExecutor executor = new ScheduledThreadPoolExecutor(4);
     final RandomSeederThread randomSeederThread =
-        new RandomSeederThread(SECURE_RANDOM_SEED_GENERATOR);
+        new RandomSeederThread(DEFAULT_INSTANCE);
     final ReseedingSplittableRandomAdapter random = ReseedingSplittableRandomAdapter
-        .getInstance(randomSeederThread, SECURE_RANDOM_SEED_GENERATOR);
+        .getInstance(randomSeederThread, DEFAULT_INSTANCE);
     for (i = 0; i < 20; i++) {
       executor.submit(() -> {
         final List<String> deck = deckCopies.get();